pyproject.toml
[project]
name = "coati-protli"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "biopython>=1.86",
    "matplotlib>=3.10.8",
    "numpy>=2.4.1",
    "pandas>=2.3.3",
    "rdkit>=2025.9.3",
    "torch>=2.9.1",
    "torchvision>=0.24.1",
]
